ledger-tool: verify: add --verify-slot-hashes
This adds:

    --verify-slot-hashes <FILENAME>
        If the file does not exist, write the slot hashes to this file;
        if the file does exist, verify slot hashes against this file.

The first case can be used to dump a list of (slot, hash) to a json file
during a replay. The second case can be used to check slot hashes against
previously recorded values.

This is useful for debugging consensus failures, eg:

    # on good commit/branch
    ledger-tool verify --verify-slot-hashes good.json ...

    # on bad commit or potentially consensus breaking branch
    ledger-tool verify --verify-slot-hashes good.json ...

On a hash mismatch an error will be logged with the expected hash vs the
computed hash.