ledger-tool: verify: add --verify-slots and --verify-slots-details
This adds:

    --verify-slots <FILENAME>
        If the file does not exist, write the slot hashes to this file;
        if the file does exist, verify slot hashes against this file.

    --verify-slots-details none|bank|tx|bank-tx
	Store the bank (=accounts) and/or transaction batches in the
	json file. For each tx batch, the bank hash is included.

The first case can be used to dump a list of (slot, hash) to a json file
during a replay. The second case can be used to check slot hashes against
previously recorded values.

This is useful for debugging consensus failures, eg:

    # on good commit/branch
    ledger-tool verify --verify-slots good.json --verify-slots-details=tx

    # on bad commit or potentially consensus breaking branch
    ledger-tool verify --verify-slots good.json --verify-slots-details=tx

On a hash mismatch an error will be logged with the expected hash vs the
computed hash.